Deleted IssuePage component and replaced it with IssueList component
