+++ src/components/IssueList.js
@@ -1,7 +1,31 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
+import styled from 'styled-components';
+import IssueCard from './IssueCard';
 
-function IssueList() {
-  return <div></div>;
+function IssuePage() {
+  const [issuesList, setIssuesList] = useState([]);
+
+  useEffect(() => {
+    axios
+      .get('https://co-make-tt-33.herokuapp.com/api/issues')
+      .then((res) => {
+        console.log('GET ISSUES', res);
+        // setIssuesList(res.data);
+      })
+      .catch((err) => {
+        console.log(err.message);
+      });
+  }, []);
+  return (
+    <IssueList>
+      {issuesList.map((issue) => {
+        return <IssueCard key={issue.id} issue={issue} />;
+      })}
+    </IssueList>
+  );
 }
 
-export default IssueList;
+export default IssuePage;
+
+const IssueList = styled.div``;
